backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
tensorflow==2.18.0
pillow==10.4.0
numpy==2.0.2
opencv-python-headless==4.10.0.84
